requirements.txt
datasets>=2.19.0
fire>=0.6.0
fschat>=0.2.36
huggingface-hub>=0.22.2
numpy>=1.26.4
openai>=1.23.3
pandas>=2.2.2
peft>=0.11.1
Pillow>=10.3.0
pydantic>=2.7.0
pyyaml>=6.0.1
scikit-learn>=1.4.2
tenacity>=8.4.1
termcolor>=2.4.0
torch>=2.8.0,<2.9
tqdm>=4.66.2
transformers>=4.57.0,<5.0.0
vllm==0.11.0  # required by WildGuard and other vLLM-based classifiers
